separate model from service
